fixed banner to display application name + version from gradle, merged test classes to improve performance
