+++ build.gradle
@@ -95,6 +95,14 @@ dependencyManagement {
   }
 }
 
+bootJar {
+  mainClassName = 'com.myhome.MyHomeServiceApplication'
+  manifest {
+    attributes('Implementation-Title': 'myhome-service',
+            'Implementation-Version': version)
+  }
+}
+
 compileJava {
   // Mapstruct options
   options.annotationProcessorPath = configurations.annotationProcessor
