+++ src/main/java/com/myhome/bootstrap/TestDataConstants.java
@@ -24,11 +24,4 @@ public class TestDataConstants {
   static final String COMMUNITY_ID = "default-community-id-for-testing";
   static final String COMMUNITY_NAME = "MyHome default community";
   static final String COMMUNITY_DISTRICT = "MyHome default community district";
-
-  static final String TEST_USER_NAME = "Test User %d";
-  static final String TEST_USER_EMAIL_ID = "testuser%d@myhome.com";
-  static final String TEST_USER_PASSWORD = "testtest";
-  static final String TEST_COMMUNITY_NAME = "Test Community %d";
-  static final String TEST_COMMUNITY_DISTRICT = "Wonderland";
-  static final String TEST_HOUSE_NAME = "Test House %d";
 }
