+++ src/main/resources/application.yml
@@ -27,7 +27,6 @@ server:
 spring:
   application:
     name: "myhome-service"
-    version: 1.2.0-release
   devtools:
     restart:
       enabled: true
