+++ src/test/java/com/myhome/controllers/integration/ControllerIntegrationTestBase.java
@@ -15,6 +15,8 @@
  */
 package com.myhome.controllers.integration;
 
+import static java.util.Optional.ofNullable;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.myhome.MyHomeServiceApplication;
@@ -24,6 +26,7 @@ import java.net.URISyntaxException;
 import java.util.Objects;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.client.TestRestTemplate;
@@ -33,17 +36,21 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import static java.util.Optional.ofNullable;
+import org.springframework.test.web.servlet.MockMvc;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT, classes = {
     MyHomeServiceApplication.class})
+@AutoConfigureMockMvc
+@ActiveProfiles(profiles = "test")
 public class ControllerIntegrationTestBase {
 
   @Autowired
   ObjectMapper objectMapper;
+  @Autowired
+  MockMvc mockMvc;
   @Autowired
   private TestRestTemplate client;
   @LocalServerPort
