--- src/test/java/com/myhome/controllers/integration/HealthCheckControllerIntegrationTest.java
@@ -1,45 +1,0 @@
-/*
- * Copyright 2020 Prathab Murugan
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.myhome.controllers.integration;
-
-import com.myhome.controllers.HealthCheckController;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.hamcrest.Matchers.containsString;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@ActiveProfiles(profiles = "test")
-class HealthCheckControllerIntegrationTest {
-
-  @Autowired MockMvc mockMvc;
-
-  @Test
-  void statusCheck() throws Exception {
-    mockMvc.perform(get("/status"))
-        .andExpect(status().isOk())
-        .andExpect(content().string(containsString(HealthCheckController.STATUS_HEALTHY)));
-  }
-}
